Abstract the watchlist processing functionality away from the controller
